Turn ResourceInformationCollector into interface
